Further fixes
