+++ packages/atlas/src/components/NftCarousel/components/MarketplaceCarouselCard/MarketplaceCarouselCard.styles.ts
@@ -2,19 +2,18 @@ import styled from '@emotion/styled'
 
 import { cVar, media, sizes } from '@/styles'
 
-export const InformationContainer = styled.div`
+export const InformationContainer = styled.div<{ isPaused: boolean }>`
   width: fit-content;
   display: flex;
   align-items: end;
   transition: all ${cVar('animationTransitionMedium')};
   margin-top: ${sizes(4)};
-  margin-bottom: ${sizes(8)};
   z-index: 12;
 
   ${media.sm} {
     position: absolute;
     margin: 0;
-    opacity: 0.25;
+    opacity: ${({ isPaused }) => (isPaused ? 1 : 0.25)};
     inset: auto auto 40px 32px;
   }
 `
