+++ packages/atlas/src/components/NftCarousel/components/MarketplaceCarouselCard/NftCarouselDetails.tsx
@@ -1,5 +1,5 @@
 import BN from 'bn.js'
-import { useLayoutEffect, useState } from 'react'
+import { useLayoutEffect, useMemo, useState } from 'react'
 import { useNavigate } from 'react-router'
 import { CSSTransition } from 'react-transition-group'
 
@@ -36,65 +36,87 @@ export const NftCarouselDetails = ({
 }) => {
   const smMatch = useMediaMatch('sm')
   const navigate = useNavigate()
+  const [timeLeft, setTimeLeft] = useState<string | null>(null)
+  const [isPaused, setIsPaused] = useState(!active)
+  const { convertBlockToMsTimestamp } = useBlockTimeEstimation()
+  const nftStatus = getNftStatus(nft, nft?.video)
+
   const creatorAvatarUrl =
     nft.owner.__typename === 'NftOwnerChannel'
       ? nft.owner.channel.avatarPhoto?.resolvedUrl
       : getMemberAvatar(nft.owner.member).url
-
-  const [timeLeft, setTimeLeft] = useState<string | null>(null)
   const thumbnailUrl = nft.video.thumbnailPhoto?.resolvedUrl
-  const { convertBlockToMsTimestamp } = useBlockTimeEstimation()
-
   const mediaUrl = nft.video.media?.resolvedUrl
-
-  const isLoading = !thumbnailUrl || !mediaUrl
-
-  const nftStatus = getNftStatus(nft, nft?.video)
   const plannedEndDateBlockTimestamp =
     nftStatus?.status === 'auction' &&
     nftStatus.auctionPlannedEndBlock &&
     convertBlockToMsTimestamp(nftStatus.auctionPlannedEndBlock)
-
+  const isLoading = !thumbnailUrl || !mediaUrl
   const name = nft.owner.__typename === 'NftOwnerChannel' ? nft.video.channel.title : nft.owner.member.handle
+  const owner = useMemo(
+    () =>
+      nft?.owner.__typename === 'NftOwnerChannel'
+        ? {
+            name,
+            assetUrl: creatorAvatarUrl,
+            onClick: () => navigate(absoluteRoutes.viewer.channel(nft.video.channel.id)),
+          }
+        : nft?.owner.__typename === 'NftOwnerMember'
+        ? {
+            name,
+            assetUrl: creatorAvatarUrl,
+            onClick: () => name && navigate(absoluteRoutes.viewer.member(name)),
+          }
+        : undefined,
+    [creatorAvatarUrl, name, navigate, nft]
+  )
 
-  const owner =
-    nft?.owner.__typename === 'NftOwnerChannel'
-      ? {
-          name,
-          assetUrl: creatorAvatarUrl,
-          onClick: () => navigate(absoluteRoutes.viewer.channel(nft.video.channel.id)),
-        }
-      : nft?.owner.__typename === 'NftOwnerMember'
-      ? {
-          name,
-          assetUrl: creatorAvatarUrl,
-          onClick: () => name && navigate(absoluteRoutes.viewer.member(name)),
-        }
-      : undefined
+  const nftDetails = useMemo(
+    () => ({
+      buyNow:
+        nftStatus?.status === 'auction' || nftStatus?.status === 'buy-now'
+          ? nftStatus.buyNowPrice
+            ? hapiBnToTokenNumber(nftStatus.buyNowPrice)
+            : undefined
+          : undefined,
+      creator: {
+        name: nft?.video.channel.title || undefined,
+        assetUrl: creatorAvatarUrl,
+        onClick: () => navigate(absoluteRoutes.viewer.channel(nft?.video.channel.id)),
+      },
+      title: nft.video.title,
+      type: 'nft',
+      topBid:
+        nftStatus?.status === 'auction' && nftStatus.topBid?.amount
+          ? hapiBnToTokenNumber(new BN(nftStatus.topBid?.amount))
+          : undefined,
+      minBid:
+        nftStatus?.status === 'auction' && nftStatus.startingPrice
+          ? hapiBnToTokenNumber(nftStatus.startingPrice)
+          : undefined,
+    }),
+    [creatorAvatarUrl, navigate, nft, nftStatus]
+  )
 
-  const nftDetails = {
-    buyNow:
-      nftStatus?.status === 'auction' || nftStatus?.status === 'buy-now'
-        ? nftStatus.buyNowPrice
-          ? hapiBnToTokenNumber(nftStatus.buyNowPrice)
-          : undefined
-        : undefined,
-    creator: {
-      name: nft?.video.channel.title || undefined,
-      assetUrl: creatorAvatarUrl,
-      onClick: () => navigate(absoluteRoutes.viewer.channel(nft?.video.channel.id)),
-    },
-    title: nft.video.title,
-    type: 'nft',
-    topBid:
-      nftStatus?.status === 'auction' && nftStatus.topBid?.amount
-        ? hapiBnToTokenNumber(new BN(nftStatus.topBid?.amount))
-        : undefined,
-    minBid:
-      nftStatus?.status === 'auction' && nftStatus.startingPrice
-        ? hapiBnToTokenNumber(nftStatus.startingPrice)
-        : undefined,
-  }
+  const avatars = useMemo(
+    () => [
+      {
+        url: nftDetails.creator?.assetUrl,
+        tooltipText: `Creator: ${nftDetails.creator?.name}`,
+        onClick: nftDetails.creator?.onClick,
+      },
+      ...(owner
+        ? [
+            {
+              url: owner?.assetUrl,
+              tooltipText: `Owner: ${owner?.name}`,
+              onClick: owner?.onClick,
+            },
+          ]
+        : []),
+    ],
+    [nftDetails.creator?.assetUrl, nftDetails.creator?.name, nftDetails.creator?.onClick, owner]
+  )
 
   useLayoutEffect(() => {
     if (plannedEndDateBlockTimestamp) {
@@ -114,7 +136,7 @@ export const NftCarouselDetails = ({
         setTimeLeft(
           `${String(hours).padStart(2, '0')}:${String(minutes).padStart(2, '0')}:${String(seconds).padStart(2, '0')}`
         )
-      }, 1000)
+      }, 12000)
 
       return () => {
         clearInterval(interval)
@@ -139,7 +161,8 @@ export const NftCarouselDetails = ({
           autoPlay={active}
           playing={active}
           muted={true}
-          onPause={(e) => (e.currentTarget.currentTime = 0)}
+          onPause={() => setIsPaused(true)}
+          onPlay={() => setIsPaused(false)}
           preload="auto"
           src={mediaUrl ?? undefined}
           poster={thumbnailUrl ?? undefined}
@@ -150,28 +173,9 @@ export const NftCarouselDetails = ({
       </VideoContainer>
       {active && (
         <CSSTransition in={active} timeout={100} classNames={transitions.names.fade} unmountOnExit>
-          <InformationContainer>
+          <InformationContainer isPaused={isPaused}>
             <DetailsContainer>
-              <AvatarGroup
-                spreadAvatars={!smMatch}
-                avatarStrokeColor="transparent"
-                avatars={[
-                  {
-                    url: nftDetails.creator?.assetUrl,
-                    tooltipText: `Creator: ${nftDetails.creator?.name}`,
-                    onClick: nftDetails.creator?.onClick,
-                  },
-                  ...(owner
-                    ? [
-                        {
-                          url: owner?.assetUrl,
-                          tooltipText: `Owner: ${owner?.name}`,
-                          onClick: owner?.onClick,
-                        },
-                      ]
-                    : []),
-                ]}
-              />
+              <AvatarGroup spreadAvatars={!smMatch} avatarStrokeColor="transparent" avatars={avatars} />
               <Text variant={smMatch ? 'h500' : 'h400'} as={smMatch ? 'h5' : 'h4'}>
                 {nftDetails.title}
               </Text>
